requirements.txt
google-adk[database]
yfinance
psutil
litellm
google-generativeai
python-dotenv
deprecated
pymysql